+++ test/helpers/mock_worksheet_action_context.dart
@@ -1,5 +1,6 @@
 import 'package:worksheet/src/core/data/formula_reference_adjuster.dart';
 import 'package:worksheet/src/core/data/worksheet_data.dart';
+import 'package:worksheet/src/core/geometry/layout_solver.dart';
 import 'package:worksheet/src/core/models/cell_coordinate.dart';
 import 'package:worksheet/src/core/models/cell_range.dart';
 import 'package:worksheet/src/interaction/clipboard/clipboard_handler.dart';
@@ -29,6 +30,8 @@ class MockWorksheetActionContext extends WorksheetActionContext {
   FormulaReferenceAdjuster? formulaReferenceAdjuster =
       defaultFormulaReferenceAdjuster;
   @override
+  LayoutSolver? layoutSolver;
+  @override
   UndoManager? undoManager;
 
   @override
@@ -47,6 +50,7 @@ class MockWorksheetActionContext extends WorksheetActionContext {
     this.onEditCell,
     this.editController,
     this.formulaReferenceAdjuster = defaultFormulaReferenceAdjuster,
+    this.layoutSolver,
     this.undoManager,
   });